+++ src/main/kotlin/com/nexsabre/hardwarereservationtool/cmd/main.kt
@@ -9,12 +9,7 @@ import com.github.ajalt.clikt.parameters.options.default
 import com.github.ajalt.clikt.parameters.options.option
 import com.github.ajalt.clikt.parameters.options.switch
 import com.github.ajalt.clikt.parameters.types.int
-import com.nexsabre.hardwarereservationtool.cmd.helpers.checkMachineExistsById
-import com.nexsabre.hardwarereservationtool.cmd.helpers.getAllMachines
-import com.nexsabre.hardwarereservationtool.cmd.helpers.getMachineById
-import com.nexsabre.hardwarereservationtool.cmd.helpers.prettyPrint
 import com.nexsabre.hardwarereservationtool.cmd.models.Hart
-import kotlin.system.exitProcess
 
 
 class HartCmd : CliktCommand() {
@@ -56,19 +51,6 @@ class Machines : CliktCommand() {
     }
 }
 
-fun showMachines(id: Int) {
-    if (id < 0) {
-        getAllMachines().prettyPrint()
-        exitProcess(0)
-    }
-    val check = checkMachineExistsById(id)
-    if (!check) {
-        println("Machine $id does not exist. Please provide a proper machine id.")
-        exitProcess(0)
-    }
-    getMachineById(id).prettyPrint()
-}
-
 fun main(array: Array<String>) {
     checkServiceIsAlive()
     HartCmd()
@@ -77,12 +59,6 @@ fun main(array: Array<String>) {
                     Machines()
             )
             .main(array)
-}
-
-fun checkServiceIsAlive() {
-    if (!Hart().isAlive()) {
-        println("Service is not available or cmd's app can not connect to it.")
-    }
 }
 
 //* main
